# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This a course project at HCMUS.",
        "title": "Swagger bookingdotcom API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/divisions": {
            "get": {
                "description": "get divisions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "divisions"
                ],
                "summary": "List vietnam divisions",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecase.DivisionAPIUseCaseFindResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/stays": {
            "get": {
                "description": "get stays example provinceCode=79\u0026districtCode=770\u0026wardCode=27127\u0026guests=2\u0026guests=0\u0026checkTimes=2024-01-31T23:34:25.191Z\u0026checkTimes=2024-09-21T06:10:36.275Z",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stays"
                ],
                "summary": "List stays",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "adults",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "CheckTimes is a pointer to an array of 2 types.DateTime, which is also weird.\nfirst element is checkin, second is checkout",
                        "name": "checkTimes",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "districtCode",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Guests is a pointer to an array of 2 int64s, which is a bit weird.\nfirst element is adults, second is children",
                        "name": "guests",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "provinceCode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "wardCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecase.StayAPIUseCaseFindStaysResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/stays/{id}": {
            "get": {
                "description": "get stay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stays"
                ],
                "summary": "Show a stay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "adults",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "CheckTimes is a pointer to an array of 2 types.DateTime, which is also weird.\nfirst element is checkin, second is checkout",
                        "name": "checkTimes",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "districtCode",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Guests is a pointer to an array of 2 int64s, which is a bit weird.\nfirst element is adults, second is children",
                        "name": "guests",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "provinceCode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "wardCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecase.StayAPIUseCaseReserveRoomResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/stays/{id}/reserve": {
            "patch": {
                "description": "reserve room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stays"
                ],
                "summary": "Reserve a room in a stay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.StayAPIUseCaseReserveRoomBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecase.StayAPIUseCaseReserveRoomResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "model.Image": {
            "type": "object",
            "properties": {
                "altText": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                }
            }
        },
        "model.ReservedTime": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "from": {
                    "type": "string",
                    "example": "2021-05-01T00:00:00Z"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "receiveTime": {
                    "type": "string",
                    "example": "2021-05-01T00:00:00Z"
                },
                "to": {
                    "type": "string",
                    "example": "2021-05-01T00:00:00Z"
                }
            }
        },
        "model.Room": {
            "type": "object",
            "properties": {
                "bedsCount": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "maxAdultGuests": {
                    "type": "integer"
                },
                "maxChildrenGuests": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pricePerDay": {
                    "type": "number"
                },
                "pricePerHour": {
                    "type": "number"
                },
                "reservedTimes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ReservedTime"
                    }
                }
            }
        },
        "model.Stay": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "districtCode": {
                    "type": "integer"
                },
                "districtName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "name": {
                    "type": "string"
                },
                "provinceCode": {
                    "type": "integer"
                },
                "provinceName": {
                    "type": "string"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Room"
                    }
                },
                "stayType": {
                    "type": "string"
                },
                "wardCode": {
                    "type": "integer"
                },
                "wardName": {
                    "type": "string"
                }
            }
        },
        "usecase.DivisionAPIUseCaseFindResult": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vnprovince.Division"
                    }
                }
            }
        },
        "usecase.StayAPIUseCaseFindStaysResult": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Stay"
                    }
                }
            }
        },
        "usecase.StayAPIUseCaseReserveRoomBody": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "from": {
                    "type": "string",
                    "example": "2021-05-01T00:00:00Z"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "receiveTime": {
                    "type": "string",
                    "example": "2021-05-01T00:00:00Z"
                },
                "roomCode": {
                    "type": "string"
                },
                "to": {
                    "type": "string",
                    "example": "2021-05-01T00:00:00Z"
                }
            }
        },
        "usecase.StayAPIUseCaseReserveRoomResult": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "districtCode": {
                    "type": "integer"
                },
                "districtName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "name": {
                    "type": "string"
                },
                "provinceCode": {
                    "type": "integer"
                },
                "provinceName": {
                    "type": "string"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Room"
                    }
                },
                "stayType": {
                    "type": "string"
                },
                "wardCode": {
                    "type": "integer"
                },
                "wardName": {
                    "type": "string"
                }
            }
        },
        "vnprovince.Division": {
            "type": "object",
            "properties": {
                "districtCode": {
                    "type": "integer"
                },
                "districtName": {
                    "type": "string"
                },
                "provinceCode": {
                    "type": "integer"
                },
                "provinceName": {
                    "type": "string"
                },
                "wardCode": {
                    "type": "integer"
                },
                "wardName": {
                    "type": "string"
                }
            }
        }
    }
}